added random candidates
